Fix onConfirm being called with undefined when selected
This changes the state of the autocomplete component to remember the
options and the selected value so that these can be used as an argument
to the onConfirm callback when the input is blurred.

Prior to this most of the onConfirm callbacks were returning undefined
once a value was selected.

This meant that it wasn't practical to have a select element reset to an empty
value when the input is cleared. This was because there wasn't any well
of telling whether a user had inputted an empty string of if the
selection was forgotten.

This still leaves a problem for select boxes that don't have an empty
attribute as you can blur the accessible autocomplete and be left in a
state where you have a select with a value whereas the autocomplete is
blank.