Split focused into inputFocused and optionFocused
This takes the single state value of focused and splits it into two
variables so that the code is easier to understand.

Previously you needed to understand that a focused of -1 meant an input
was focused and that a focused of null meant nothing was focused. This
was rather confusing.